+++ gtests/rint/rint_callback.cc
@@ -54,12 +54,12 @@ double getFuncOp(double *data) {
   return LIBM_FUNC(rint)(data[0]);
 }
 
-float getExpected(float *data) {
+double getExpected(float *data) {
   auto val = alm_mp_rintf(data[0]);
   return val;
 }
 
-double getExpected(double *data) {
+long double getExpected(double *data) {
   auto val = alm_mp_rint(data[0]);
   return val;
 }
